add the management user data and add aws provider to everything
This adds the aws provider block to the head and compute nodes, the
flavor of the node still needs to be changed to the HPC variety of
node though. Also making sure the nodes can access the mgmt server
by splitting the userdata scripts for both types of systems. This
does hard code the mgmt server though I'm not sure how to fix this
as its part of the bootstrap process.